requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
beautifulsoup4==4.12.2; python_full_version >= '3.6.0'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
crispy-bootstrap5==2023.10; python_version >= '3.8'
cryptography==41.0.5
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.2.7; python_version >= '3.8'
django-allauth==0.58.2; python_version >= '3.7'
django-bootstrap-v5==1.0.11; python_version >= '3.6' and python_version < '4.0'
django-crispy-forms==2.1; python_version >= '3.8'
django-htmx==1.17.2; python_version >= '3.8'
django-widget-tweaks==1.5.0; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
numpy==1.26.2; python_version == '3.11'
oauthlib==3.2.2; python_version >= '3.6'
pandas==2.1.3; python_version >= '3.9'
pillow==10.1.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pycparser==2.21
pyjwt[crypto]==2.8.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
tzdata==2023.3; python_version >= '2'
urllib3==2.1.0; python_version >= '3.8'


